+++ src/main/java/com/swyp/index/application/report/CalendarService.java
@@ -1,4 +1,4 @@
-package com.swyp.index.application.calendar;
+package com.swyp.index.application.report;
 
 import com.swyp.index.domain.bookshelf.Bookshelf;
 import com.swyp.index.domain.bookshelf.PageRecord;
@@ -6,7 +6,7 @@ import com.swyp.index.domain.user.User;
 import com.swyp.index.infrastructure.repository.BookshelfRepository;
 import com.swyp.index.infrastructure.repository.PageRecordRepository;
 import com.swyp.index.infrastructure.repository.UserRepository;
-import com.swyp.index.presentation.dto.calendar.CalendarResponse;
+import com.swyp.index.presentation.dto.report.CalendarResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
